Online Order Progress 3.5 hours

+++ week04/OnlineOrdering/Product.cs
@@ -15,7 +15,7 @@ public class Product
         _quantity = quantity;
     }
     //Methods
-    public double ComputeSum()
+    public float ComputeSum()
     {
         return _price * _quantity;
     }
